refactor: исправить проблемы с checkstyle
