+++ src/handlers/EpicsHandler.java
@@ -8,7 +8,6 @@ import exceptions.TaskException;
 import servers.HttpTaskServer;
 import tasks.EpicTask;
 import tasks.SubTask;
-import test.LocalDateTypeAdapter;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
