+++ src/handlers/LocalDateTypeAdapter.java
@@ -1,4 +1,4 @@
-package test;
+package handlers;
 
 import com.google.gson.*;
