+++ src/handlers/PrioritizedHandler.java
@@ -6,7 +6,6 @@ import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import servers.HttpTaskServer;
 import tasks.Task;
-import test.LocalDateTypeAdapter;
 
 import java.io.IOException;
 import java.time.LocalDate;
