+++ src/handlers/SubtasksHandler.java
@@ -7,7 +7,6 @@ import com.sun.net.httpserver.HttpHandler;
 import exceptions.TaskException;
 import servers.HttpTaskServer;
 import tasks.SubTask;
-import test.LocalDateTypeAdapter;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
